Bug 704216: Fix prototype for gdk_quartz_device_core_warp.

+++ gdk/quartz/gdkdevice-core-quartz.c
@@ -52,8 +52,8 @@ static void gdk_quartz_device_core_set_window_cursor (GdkDevice *device,
                                                       GdkCursor *cursor);
 static void gdk_quartz_device_core_warp (GdkDevice *device,
                                          GdkScreen *screen,
-                                         gint       x,
-                                         gint       y);
+                                         gdouble    x,
+                                         gdouble    y);
 static void gdk_quartz_device_core_query_state (GdkDevice        *device,
                                                 GdkWindow        *window,
                                                 GdkWindow       **root_window,
